create amoba.cpp
I've written this piece of code one week ago, but I could't continue it until now.
So this is my first really commit, and it seems to me, that I'm out of time now.
Am I crazy or brave? I don't know, but "the easy day was yesterday".
Let's get started!